pyproject.toml
[tool.poetry]
name = "ferm_ge"
version = "0.1.0"
description = "A Fair Empirical Risk Minimization with Generalized Entropy"
authors = []

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
numpy = ">=1.22,<1.24"
scikit-learn = "^1.2.0"
matplotlib = "^3.6.2"
pandas = "^1.5.2"
requests = "^2.28.1"
numba = "^0.56.4"
pyyaml = "^6.0"
psutil = "^5.9.4"

[tool.poetry.group.dev.dependencies]
black = {extras = ["jupyter"], version = "^22.12.0"}
mypy = "^0.991"
ipython = "8.7.0"
ipywidgets = "7.7.1"
types-requests = "^2.28.11.7"
types-pyyaml = "^6.0.12.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80

[tool.mypy]
follow_imports = "normal"
ignore_missing_imports = true
check_untyped_defs = true
warn_return_any = true
pretty = false
show_column_numbers = true
plugins = "numpy.typing.mypy_plugin"
